+++ src/main/java/io/unthrottled/doki/settings/ThemeSettingsUI.java
@@ -10,7 +10,6 @@ import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.openapi.util.NlsContexts;
 import com.intellij.ui.ColorUtil;
 import com.intellij.ui.JBColor;
-import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.UIUtil;
 import io.unthrottled.doki.config.ThemeConfig;
 import io.unthrottled.doki.stickers.CurrentSticker;
@@ -26,7 +25,7 @@ public class ThemeSettingsUI implements SearchableConfigurable, Configurable.NoS
   private final ThemeSettingsModel themeSettingsModel = ThemeSettings.createThemeSettingsModel();
   private ThemeSettingsModel initialThemeSettingsModel = ThemeSettings.createThemeSettingsModel();
 
-  private JTabbedPane tabbedPane1;
+  private JTabbedPane fontTabPane;
   private JPanel rootPane;
   private JComboBox currentThemeWomboComboBox;
   private JCheckBox showStickerCheckBox;
@@ -42,11 +41,13 @@ public class ThemeSettingsUI implements SearchableConfigurable, Configurable.NoS
   private JButton chooseImageButton;
   private JCheckBox useCustomStickerCheckBox;
   private JTextPane generalLinks;
-  private JSpinner customFontSize;
-  private JCheckBox overrideEditorFontSizeCheckBox;
   private JSlider notificationOpacitySlider;
   private JCheckBox makeNotificationsTransparentCheckBox;
   private JCheckBox themeChangeAnimationCheckBox;
+  private JSpinner customFontSize;
+  private JCheckBox overrideEditorFontSizeCheckBox;
+  private JComboBox consoleFontWomboComboBox;
+  private JCheckBox overrideConsoleFont;
 
 
   @Override
@@ -137,6 +138,10 @@ public class ThemeSettingsUI implements SearchableConfigurable, Configurable.NoS
     overrideEditorFontSizeCheckBox.addActionListener(e ->
       themeSettingsModel.setCustomFontSize(overrideEditorFontSizeCheckBox.isSelected()));
 
+    overrideConsoleFont.setSelected(initialThemeSettingsModel.isOverrideConsoleFont());
+    overrideConsoleFont.addActionListener(e ->
+      themeSettingsModel.setOverrideConsoleFont(overrideConsoleFont.isSelected()));
+
     SpinnerNumberModel customFontSizeModel = new SpinnerNumberModel(
       initialThemeSettingsModel.getCustomFontSizeValue(),
       1,
@@ -220,6 +225,12 @@ public class ThemeSettingsUI implements SearchableConfigurable, Configurable.NoS
     });
 
     currentThemeWomboComboBox = ThemeSettings.INSTANCE.createThemeComboBoxModel(
+      () -> this.themeSettingsModel == null ?
+        ThemeSettings.createThemeSettingsModel() :
+        themeSettingsModel
+    );
+
+    consoleFontWomboComboBox = ThemeSettings.INSTANCE.createConsoleFontComboBoxModel(
       () -> this.themeSettingsModel == null ?
         ThemeSettings.createThemeSettingsModel() :
         themeSettingsModel