+++ src/main/kotlin/io/unthrottled/doki/notification/UpdateNotification.kt
@@ -84,14 +84,7 @@ private fun buildUpdateMessage(
     <body>
     <h2>What's New?</h2>
     <ul>
-        <li>3 NekoPara Themes!
-            <ul>
-                <li>Chocola</li>
-                <li>Vanilla</li>
-                <li>Coconut</li>
-            </ul>
-        </li>
-        <li>Better 2021.2 Build Support.</li>
+        <li>Added global console font config.</li>
     </ul>
     Please see the <a href="https://github.com/doki-theme/doki-theme-jetbrains/blob/master/changelog/CHANGELOG.md">
         changelog</a> for more details.